Fixes for amd64.  Seems to work ok.
git-svn-id: http://svn.mythtv.org/svn/trunk@6815 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mfd/clientlib/discoverythread.cpp
@@ -310,7 +310,7 @@ void DiscoveryThread::run()
                 mdnsd_in(
                             mdns_daemon,
                             &m,
-                            (unsigned long int)from.sin_addr.s_addr,
+                            (unsigned int)from.sin_addr.s_addr,
                             from.sin_port
                         );
             }
@@ -359,7 +359,7 @@ void DiscoveryThread::run()
                 mdnsd_in(
                             mdns_ipaddr_daemon,
                             &m,
-                            (unsigned long int)from.sin_addr.s_addr,
+                            (unsigned int)from.sin_addr.s_addr,
                             from.sin_port
                         );
             }
